ci: Install code without editable install
For completness, install the code as a package undert site-packages
and not as an editable install. This should have no difference in
almost all cases as installs in CI are by definition testing the
same code as in the state of the repository, but it at least is
testing the install location a user would have.